move IsLastNamespaceSegmentTheSame from public
