+++ packages/http-client-csharp/generator/Microsoft.TypeSpec.Generator.Input/src/Extensions/StringExtensions.cs
@@ -1,7 +1,6 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-using System;
 using System.Diagnostics.CodeAnalysis;
 using System.Text;
 using Microsoft.CodeAnalysis.CSharp;
@@ -79,50 +78,5 @@ namespace Microsoft.TypeSpec.Generator.Input.Extensions
 
         [return: NotNullIfNotNull(nameof(name))]
         public static string ToVariableName(this string name) => name.ToIdentifierName(isCamelCase: false);
-
-        /// <summary>
-        /// Checks if two namespaces share the same last segment
-        /// </summary>
-        /// <param name="left">the first namespace</param>
-        /// <param name="right">the second namespace</param>
-        /// <returns></returns>
-        public static bool IsLastNamespaceSegmentTheSame(string left, string right)
-        {
-            // finish this via Span API
-            var leftSpan = left.AsSpan();
-            var rightSpan = right.AsSpan();
-            // swap if left is longer, we ensure left is the shorter one
-            if (leftSpan.Length > rightSpan.Length)
-            {
-                var temp = leftSpan;
-                leftSpan = rightSpan;
-                rightSpan = temp;
-            }
-            for (int i = 1; i <= leftSpan.Length; i++)
-            {
-                var lc = leftSpan[^i];
-                var rc = rightSpan[^i];
-                // check if each char is the same from the right-most side
-                // if both of them are dot, we finished scanning the last segment - and if we could be here, meaning all of them are the same, return true.
-                if (lc == '.' && rc == '.')
-                {
-                    return true;
-                }
-                // if these are different - there is one different character, return false.
-                if (lc != rc)
-                {
-                    return false;
-                }
-            }
-
-            // we come here because we run out of characters in left - which means left does not have a dot.
-            // if they have the same length, they are identical, return true
-            if (leftSpan.Length == rightSpan.Length)
-            {
-                return true;
-            }
-            // otherwise, right is longer, we check its next character, if it is the dot, return true, otherwise return false.
-            return rightSpan[^(leftSpan.Length + 1)] == '.';
-        }
     }
 }
