+++ packages/http-client-csharp/generator/Microsoft.TypeSpec.Generator/test/Utilities/StringExtensionsTests.cs
@@ -3,9 +3,7 @@
 
 using System;
 using System.Collections.Generic;
-using PublicStringExtensions = Microsoft.TypeSpec.Generator.Input.Extensions.StringExtensions;
 using Microsoft.TypeSpec.Generator.Utilities;
-using InternalStringExtensions = Microsoft.TypeSpec.Generator.Utilities.StringExtensions;
 using NUnit.Framework;
 
 namespace Microsoft.TypeSpec.Generator.Tests.Utilities
@@ -116,7 +114,7 @@ namespace Microsoft.TypeSpec.Generator.Tests.Utilities
         [TestCase("yield", true)]
         public void TestIsCSharpKeyword(string name, bool isKeyword)
         {
-            var result = InternalStringExtensions.IsCSharpKeyword(name);
+            var result = StringExtensions.IsCSharpKeyword(name);
             Assert.AreEqual(isKeyword, result);
         }
 
@@ -136,31 +134,13 @@ namespace Microsoft.TypeSpec.Generator.Tests.Utilities
         public void ValidateGetFormattableStringFormatParts(string format, IReadOnlyList<Part> parts)
         {
             var i = 0;
-            foreach (var (span, isLiteral, index) in InternalStringExtensions.GetFormattableStringFormatParts(format))
+            foreach (var (span, isLiteral, index) in StringExtensions.GetFormattableStringFormatParts(format))
             {
                 Assert.AreEqual(parts[i].Value, span.ToString());
                 Assert.AreEqual(parts[i].IsLiteral, isLiteral);
                 Assert.AreEqual(parts[i].ArgumentIndex, index);
                 i++;
             }
-        }
-
-        [TestCase("Foo", "Foo", ExpectedResult = true)]
-        [TestCase("Foo", "Bar", ExpectedResult = false)]
-        [TestCase("Foo", "_Foo", ExpectedResult = false)]
-        [TestCase("_Foo", "Foo", ExpectedResult = false)]
-        [TestCase("Foo", "Bar.Foo", ExpectedResult = true)]
-        [TestCase("Bar.Foo", "Foo", ExpectedResult = true)]
-        [TestCase("Foo", "Bar._Foo", ExpectedResult = false)]
-        [TestCase("Bar._Foo", "Foo", ExpectedResult = false)]
-        [TestCase("Foo", "/Foo", ExpectedResult = false)]
-        [TestCase("/Foo", "Foo", ExpectedResult = false)]
-        [TestCase(".Foo", ".Foo", ExpectedResult = true)]
-        [TestCase("Foo", ".Foo", ExpectedResult = true)]
-        [TestCase(".Foo", "Foo", ExpectedResult = true)]
-        public bool ValidateIsLastNamespaceSegmentTheSame(string left, string right)
-        {
-            return PublicStringExtensions.IsLastNamespaceSegmentTheSame(left, right);
         }
 
         public record Part(string Value, bool IsLiteral, int ArgumentIndex);
